feat: 예약 상태 정보 조회 기능 추가

+++ backend/src/main/java/com/metamong/mt/domain/reservation/repository/mybatis/ReservationMapper.java
@@ -10,6 +10,6 @@ import com.metamong.mt.domain.reservation.dto.response.ReservationResponseDto;
 @Repository
 @Mapper
 public interface ReservationMapper {
-	List<ReservationResponseDto> findReservationByConsId(int consId);
-	ReservationResponseDto findReservationByRvtId(int rvtId);
+	List<ReservationResponseDto> findReservationByConsId(Long consId);
+	ReservationResponseDto findReservationByRvtId(Long rvtId);
 }
