+++ backend/src/main/java/com/metamong/mt/domain/reservation/service/ReservationService.java
@@ -5,6 +5,6 @@ import java.util.List;
 import com.metamong.mt.domain.reservation.dto.response.ReservationResponseDto;
 
 public interface ReservationService {
-	List<ReservationResponseDto> findReservationByConsId(int consId);
-	ReservationResponseDto findReservationByRvtId(int rvtId);
+	List<ReservationResponseDto> findReservationByConsId(Long consId);
+	ReservationResponseDto findReservationByRvtId(Long rvtId);
 }
